ME HLE: use Register_*() pattern with separate files for ME modules
Move sceSysEventForKernel, sceSysreg_driver, and sceMeCore_driver out of
moduleList[] and into proper Register_*() functions called at the end of
RegisterAllModules(), following PPSSPP convention of one file per module:
- Core/HLE/sceSysEvent.cpp/.h (sceSysEventForKernel)
- Core/HLE/sceSysreg.cpp/.h   (sceSysreg_driver)
- Core/HLE/sceMeCore.cpp/.h   (sceMeCore_driver)

This preserves syscall numbering for savestate compatibility.
